Pull the Area chart primitive out of the LineChart component
