+++ ui/app/components/chart-primitives/area.hbs
@@ -1,0 +1,13 @@
+<defs>
+  <linearGradient x1="0" x2="0" y1="0" y2="1" class="{{this.colorClass}}" id="{{this.fillId}}">
+    <stop class="start" offset="0%" />
+    <stop class="end" offset="100%" />
+  </linearGradient>
+  <clipPath id="{{this.maskId}}">
+    <path class="fill" d="{{this.area}}" />
+  </clipPath>
+</defs>
+<g class="area {{this.colorClass}}" ...attributes>
+  <path class="line" d="{{this.line}}" />
+  <rect class="fill" x="0" y="0" width="{{@width}}" height="{{@height}}" fill="url(#{{this.fillId}})" clip-path="url(#{{this.maskId}})" />
+</g>
